still functioning with modular architecture
